requirements.txt
alembic ~= 1.16.4
authlib ~= 1.6.3
fastapi ~= 0.116.1
httpx ~= 0.28.1
itsdangerous ~= 2.2.0
jinja2 ~= 3.1.6
passlib[bcrypt] ~= 1.7.4
python-dotenv ~= 1.1.1
python-jose[cryptography] ~= 3.3.5
python-multipart ~= 0.0.20
SQLModel ~= 0.0.24
uvicorn ~= 0.35.0
yfinance ~= 0.2.65